# openapi.yaml
openapi: 3.1.0
info:
  title: media
  version: 1.0.0
  description: |
    Media service: authenticated uploads, content-addressed storage, image normalisation and
    variants, signed URLs and range-capable delivery over HTTP.

    Three distinct auth models coexist on this service:
    - `/v1/*` (except the ticketed upload below): `Authorization: Bearer <API key>` from
      `MEDIA_API_KEYS`. No role model — any valid key may call any `/v1/*` route; there is no
      per-route role gate anywhere in this service.
    - `PUT /v1/uploads/{token}`: no `Authorization` header at all. The single-use upload ticket
      in the `{token}` path segment (minted by `POST /v1/uploads`) is the only credential.
    - `GET /files/{id}/{variant}`: no `Authorization` header. A `public`-visibility file is
      served to anyone; a `private`-visibility file additionally requires a valid `exp`/`sig`
      query-string signature (HMAC-SHA256, minted by `POST /v1/files/{id}/urls` or implicitly by
      the `urls` field on any file response). Signature verification happens inside the route
      handler itself — it is not a securityScheme evaluated by auth middleware.

    Every error response shares one envelope: `{ "error": { "code", "message", "details"? } }`.
    `code` is a stable string (e.g. `NOT_FOUND`, `VALIDATION_FAILED`); `details`, when present, is
    either an object (domain-specific extra fields) or an array (Ajv validation failures, one
...
security:
  - ApiKeyAuth: []
...
paths:
...
  /v1/files:
    put:
      operationId: media.files.upload
      tags: [Files]
      summary: Upload a file
      description: |
        Raw-body upload — **not** `multipart/form-data**; no multipart dependency exists in this
        service. The request body is the file's bytes verbatim, piped straight into storage
        without buffering the whole thing in memory.

        The real type is sniffed from the leading bytes server-side (`TypeSniffer`); any
        client-supplied `Content-Type` is ignored for storage/serving decisions. Raster images
        (`image/jpeg`, `png`, `webp`, `gif`, `avif`) are re-encoded on upload by default
        (`STRIP_IMAGE_METADATA=true`), which strips EXIF/GPS/XMP/ICC metadata (ICC profile is
        re-embedded) and defeats polyglot files; `application/pdf` and, if enabled via
        `ALLOWED_TYPES`, `image/svg+xml` are stored as-is.

        Filename: the `name` query parameter wins; if absent, the `X-File-Name` header is used
        (URL-decoded if possible); if both are absent, a generic `file.<ext>` name is generated.
        Either way the name is sanitised and its extension is always re-derived from the sniffed
        type, never trusted from the client.

        **Maximum upload size**: `MAX_UPLOAD_BYTES` (default 25 MiB), enforced inside the storage
        layer's write-stream as bytes arrive. Fastify's own `bodyLimit` config on this app
        (`maxUploadBytes + 1024`) is set but is dead code for this route: its enforcement only
        applies to parsers registered with `parseAs: 'string'|'buffer'`, and this route's raw
        passthrough content-type parser bypasses that check entirely. The byte cap above is real
        and enforced — just not by the mechanism the Fastify config implies.
      security:
        - ApiKeyAuth: []
      x-stack-visibility: authenticated
      parameters:
        - $ref: '#/components/parameters/UploadVisibility'
        - $ref: '#/components/parameters/UploadName'
        - $ref: '#/components/parameters/XFileName'
      requestBody:
        required: true
        content:
          '*/*':
            schema:
              type: string
              format: binary
              description: Raw file bytes. Any Content-Type is accepted; the real type is sniffed server-side.
      responses:
        '201':
          description: File stored (or deduplicated against existing identical content).
          headers:
            Location:
              schema: { type: string }
              description: "`/v1/files/{id}` of the stored file."
          content:
            application/json:
              schema:
                type: object
                required: [file]
                properties:
                  file: { $ref: '#/components/schemas/FileView' }
        '400':
          description: |
            `VALIDATION_FAILED` (malformed `visibility`/`name` query parameter) — or, from the
            upload stream itself: `EMPTY` (zero-byte body) or `STREAM_ERROR` (the request stream
            errored or was aborted partway through).
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorEnvelope' }
        '401':
          $ref: '#/components/responses/Unauthorized'
        '413':
          $ref: '#/components/responses/TooLarge'
        '415':
          $ref: '#/components/responses/UnsupportedType'
        '422':
          $ref: '#/components/responses/InvalidImage'
        '429':
          $ref: '#/components/responses/RateLimited'

    get:
      operationId: media.files.list
      tags: [Files]
      summary: List files owned by the calling API key
      description: |
        Keyset pagination, newest first. `nextCursor` is an opaque, base64url-encoded
        `<createdAt>:<id>` token; pass it back as `cursor` to fetch the next page. `null` means no
        further pages.
      security:
        - ApiKeyAuth: []
      x-stack-visibility: authenticated
      parameters:
        - name: limit
          in: query
          required: false
          description: "Page size, 1-100. Default 20."
          schema: { type: string, pattern: '^([1-9]|[1-9][0-9]|100)$' }
        - name: cursor
          in: query
          required: false
          description: Opaque `nextCursor` value from a previous page.
          schema: { type: string, maxLength: 128 }
      responses:
        '200':
          description: A page of files owned by the calling key.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/FileList' }
        '400':
          description: |
            `VALIDATION_FAILED` (malformed `limit`/`cursor`) or `INVALID_ARGUMENT` (`cursor`
            passed schema validation but does not decode to a well-formed keyset cursor).
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorEnvelope' }
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/RateLimited'
  /v1/files/{id}:
    get:
      operationId: media.files.get
      tags: [Files]
      summary: Get a file's metadata
      security:
        - ApiKeyAuth: []
      x-stack-visibility: authenticated
      parameters:
        - $ref: '#/components/parameters/FileId'
      responses:
        '200':
          description: The file, scoped to the calling API key.
          content:
            application/json:
              schema:
                type: object
                required: [file]
                properties:
                  file: { $ref: '#/components/schemas/FileView' }
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          description: "`NOT_FOUND` — file does not exist, is soft-deleted, or is owned by a different API key."
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorEnvelope' }
        '429':
          $ref: '#/components/responses/RateLimited'

    patch:
      operationId: media.files.patch
      tags: [Files]
      summary: Update a file's visibility and/or name
      security:
        - ApiKeyAuth: []
      x-stack-visibility: authenticated
      parameters:
        - $ref: '#/components/parameters/FileId'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/FilePatchBody' }
      responses:
        '200':
          description: The updated file.
          content:
            application/json:
              schema:
                type: object
                required: [file]
                properties:
                  file: { $ref: '#/components/schemas/FileView' }
        '400':
          description: "`VALIDATION_FAILED` — empty body, or a `visibility`/`name` value outside its allowed shape."
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorEnvelope' }
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          description: "`NOT_FOUND` — file does not exist, is soft-deleted, or is owned by a different API key."
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorEnvelope' }
        '429':
          $ref: '#/components/responses/RateLimited'

    delete:
      operationId: media.files.delete
      tags: [Files]
      summary: Soft-delete a file
      description: |
        Marks the file deleted; it stops appearing in `GET /v1/files`/`GET /v1/files/{id}` and its
        delivery route immediately. Bytes are only physically purged after `DELETE_GRACE_DAYS`
        (default 7). Reversible within that window via `POST /v1/files/{id}/restore`.
      security:
        - ApiKeyAuth: []
      x-stack-visibility: authenticated
      parameters:
        - $ref: '#/components/parameters/FileId'
      responses:
        '204':
          description: Soft-deleted.
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          description: "`NOT_FOUND` — file does not exist, is already soft-deleted, or is owned by a different API key."
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorEnvelope' }
        '429':
          $ref: '#/components/responses/RateLimited'
  /v1/files/{id}/restore:
    post:
      operationId: media.files.restore
      tags: [Files]
      summary: Undo a soft delete
      description: Only works within the `DELETE_GRACE_DAYS` grace window, before bytes are physically purged.
      security:
        - ApiKeyAuth: []
      x-stack-visibility: authenticated
      parameters:
        - $ref: '#/components/parameters/FileId'
      responses:
        '200':
          description: The restored file.
          content:
            application/json:
              schema:
                type: object
                required: [file]
                properties:
                  file: { $ref: '#/components/schemas/FileView' }
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          description: "`NOT_FOUND` — file does not exist or is owned by a different API key (restorable regardless of current soft-delete state, as long as both are true)."
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorEnvelope' }
        '429':
          $ref: '#/components/responses/RateLimited'
  /v1/files/{id}/urls:
    post:
      operationId: media.files.signUrls
      tags: [Files]
      summary: (Re-)sign delivery URLs for a file
      description: |
        Returns a fresh `urls` map (same shape as `FileView.urls`) with a caller-chosen TTL. For a
        `public` file every URL is unsigned and `expiresAt` is always `null`, regardless of `ttl`.
      security:
        - ApiKeyAuth: []
      x-stack-visibility: authenticated
      parameters:
        - $ref: '#/components/parameters/FileId'
        - name: ttl
          in: query
          required: false
          description: |
            Signature lifetime in seconds. Digits only; server-side clamped to a maximum of
            604800 (7 days) even if a larger value is supplied. Defaults to `SIGNED_URL_TTL_SEC`
            (service default 900s) when omitted.
          schema: { type: string, pattern: '^\d{1,7}$' }
      responses:
        '200':
          description: Freshly signed URLs for this file.
          content:
            application/json:
              schema:
                type: object
                required: [urls]
                properties:
                  urls: { $ref: '#/components/schemas/UrlsMap' }
        '400':
          description: "`VALIDATION_FAILED` — malformed `ttl`."
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorEnvelope' }
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          description: "`NOT_FOUND` — file does not exist, is soft-deleted, or is owned by a different API key."
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorEnvelope' }
        '429':
          $ref: '#/components/responses/RateLimited'
  /v1/uploads:
    post:
      operationId: media.tickets.create
      tags: [Uploads]
      summary: Issue a single-use upload ticket
      description: |
        Lets a browser (or any caller without an API key) upload one file directly to
        `PUT /v1/uploads/{token}`. The ticket is consumed exactly once, on the first successful or
        failed upload attempt against it (a failed attempt releases it for retry); it always
        expires after `UPLOAD_TICKET_TTL_SEC` (default 900s) regardless of use.

        Note: the plaintext ticket token is forwarded, unhashed, into this action's own audit
        event (`target.id`) — unlike every other credential in this service (API keys, and the
        ticket token itself in normal use), which is only ever stored or forwarded as a hash. The
        ticket's short lifetime and single use bound the exposure, but this is a real, confirmed
        inconsistency with the rest of the platform's credential-handling discipline.
      security:
        - ApiKeyAuth: []
      x-stack-visibility: authenticated
      requestBody:
        required: false
        content:
          application/json:
            schema: { $ref: '#/components/schemas/TicketCreateBody' }
      responses:
        '201':
          description: Ticket issued.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/UploadTicket' }
        '400':
          description: |
            `VALIDATION_FAILED` (malformed body) or `INVALID_ARGUMENT` (`maxBytes` outside
            `1..MAX_UPLOAD_BYTES`, or `allowedTypes` includes a type this service does not itself
            allow via `ALLOWED_TYPES`).
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorEnvelope' }
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/RateLimited'
  /v1/uploads/{token}:
    put:
      operationId: media.files.uploadTicketed
      tags: [Uploads]
      summary: Upload against a single-use ticket
      description: |
        Ticket-only auth: **no `Authorization` header is sent or checked on this route at all** —
        the `{token}` path segment is the sole credential, matched against the hash of a live,
        unused, unexpired ticket from `POST /v1/uploads`. Designed to be called directly from a
        browser (see the `Cors` allow-list for this route's CORS behaviour).

        Same raw-body semantics as `PUT /v1/files` (not `multipart/form-data`) and the same
        server-side type sniffing, default image re-encoding/metadata stripping, and filename
        resolution (`name` query parameter, else `X-File-Name` header, else generated). Visibility
        and the upload's own byte/type limits come from the ticket itself (set when it was
        created), not from this request.

        On any failure the ticket is released (not burned), so the client may retry against the
        same `{token}` until it expires.
      security: []
      x-stack-visibility: public
      parameters:
        - name: token
          in: path
          required: true
          description: Single-use upload ticket, exactly as returned by `POST /v1/uploads`.
          schema: { type: string, pattern: '^[A-Za-z0-9_-]{43}$' }
        - $ref: '#/components/parameters/UploadName'
        - $ref: '#/components/parameters/XFileName'
      requestBody:
        required: true
        content:
          '*/*':
            schema:
              type: string
              format: binary
              description: Raw file bytes. Any Content-Type is accepted; the real type is sniffed server-side.
      responses:
        '201':
          description: File stored (or deduplicated against existing identical content).
          headers:
            Location:
              schema: { type: string }
              description: "`/v1/files/{id}` of the stored file."
          content:
            application/json:
              schema:
                type: object
                required: [file]
                properties:
                  file: { $ref: '#/components/schemas/FileView' }
        '400':
          description: |
            `VALIDATION_FAILED` (malformed `name` query parameter) — or, from the upload stream
            itself: `EMPTY` (zero-byte body) or `STREAM_ERROR` (the request stream errored or was
            aborted partway through).
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorEnvelope' }
        '401':
          description: "`INVALID_TICKET` — the ticket is unknown, already used, or expired. Distinct from the API-key `UNAUTHORIZED` used elsewhere: this route never checks `Authorization`."
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorEnvelope' }
        '413':
          description: |
            `TOO_LARGE` — upload exceeds the ticket's own `maxBytes` (which can be tighter than,
            but never looser than, the service-wide `MAX_UPLOAD_BYTES`).
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorEnvelope' }
        '415':
          $ref: '#/components/responses/UnsupportedType'
        '422':
          $ref: '#/components/responses/InvalidImage'
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: http
      scheme: bearer
      description: |
        `Authorization: Bearer <secret>`. Keys come from `MEDIA_API_KEYS` (`id:secret` pairs; this
        service has no role model — any valid key may call any `/v1/*` route). Every configured
        key is compared in constant time (SHA-256 + `timingSafeEqual`, no early exit), so timing
        cannot reveal whether, or which, key matched.
  parameters:
    FileId:
      name: id
      in: path
      required: true
      schema: { $ref: '#/components/schemas/FileId' }
    UploadVisibility:
      name: visibility
      in: query
      required: false
      description: Defaults to `private` when omitted.
      schema: { $ref: '#/components/schemas/Visibility' }
    UploadName:
      name: name
      in: query
      required: false
      description: |
        Display filename, 1-255 characters. Takes precedence over the `X-File-Name` header when
        both are given. Sanitised server-side and its extension is always re-derived from the
        sniffed type; a generic `file.<ext>` name is used if neither `name` nor `X-File-Name` is
        given.
      schema: { $ref: '#/components/schemas/FileName' }
    XFileName:
      name: X-File-Name
      in: header
      required: false
      description: |
        Fallback filename source, used only when the `name` query parameter is absent.
        URL-decoded if possible, used raw otherwise.
      schema: { type: string }
  responses:
    Unauthorized:
      description: "`UNAUTHORIZED` — missing or non-matching `Authorization: Bearer` API key."
      headers:
        WWW-Authenticate:
          schema: { type: string, enum: [Bearer] }
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
    TooLarge:
      description: |
        `TOO_LARGE` — upload exceeds `MAX_UPLOAD_BYTES` (service default 25 MiB). Enforced inside
        the storage layer's write-stream as bytes arrive, not by Fastify's `bodyLimit` config
        (which is set but inert for this raw-passthrough route).
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
    UnsupportedType:
      description: "`UNSUPPORTED_TYPE` — the sniffed type (from the file's real leading bytes, never the client-supplied `Content-Type`) is unrecognised or not in `ALLOWED_TYPES`."
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
    InvalidImage:
      description: "`INVALID_IMAGE` — the sniffed type is a raster image format, but the bytes fail to decode, have no dimensions, or exceed `MAX_IMAGE_PIXELS` (decompression-bomb guard)."
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
...
  schemas:
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code: { type: string, description: "Stable machine-readable code, e.g. `NOT_FOUND`, `VALIDATION_FAILED`." }
            message: { type: string }
            details:
              description: |
                Present only for some codes: an object of domain-specific extra fields (e.g.
                `{maxBytes}` on `TOO_LARGE`), or, for `VALIDATION_FAILED`, an array of
                `{path, message, params}` — one entry per violated Ajv constraint.
...
    FileView:
      type: object
      required: [id, name, mime, size, width, height, sha256, visibility, createdAt, urls]
      properties:
        id: { $ref: '#/components/schemas/FileId' }
        name: { type: string, description: Sanitised display name; extension always matches the sniffed type. }
        mime: { type: string, description: Server-sniffed MIME type of the `original` — never the client-supplied `Content-Type`. }
        size: { type: integer, minimum: 0, description: Byte size of the `original`. }
        width:
          type: [integer, 'null']
          description: Pixel width. `null` for a non-raster file (`application/pdf`, `image/svg+xml`); always set for a raster image.
        height:
          type: [integer, 'null']
          description: Pixel height. `null` for a non-raster file; always set for a raster image.
        sha256: { type: string, pattern: '^[0-9a-f]{64}$', description: "Content hash of the `original`'s bytes (after metadata-stripping normalisation, when applied). Drives content-addressed dedup." }
        visibility: { $ref: '#/components/schemas/Visibility' }
        createdAt: { type: string, format: date-time }
        urls: { $ref: '#/components/schemas/UrlsMap' }
    FileList:
      type: object
      required: [items, nextCursor]
      properties:
        items:
          type: array
          items: { $ref: '#/components/schemas/FileView' }
        nextCursor:
          type: [string, 'null']
          description: Opaque keyset cursor for the next page, or `null` when this is the last page.
...